all product and category APIs verified

+++ OnlineShoppingApplication/src/main/java/com/OnlineShoppingApp/Service/ProductServiceImpl.java
@@ -7,11 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.OnlineShoppingApp.Entity.Category;
+import com.OnlineShoppingApp.Entity.CurrentSession;
 import com.OnlineShoppingApp.Entity.Product;
 import com.OnlineShoppingApp.Exception.CategoryException;
 import com.OnlineShoppingApp.Exception.ProductException;
 import com.OnlineShoppingApp.Repository.CategoryDao;
 import com.OnlineShoppingApp.Repository.ProductDao;
+import com.OnlineShoppingApp.Repository.SessionDao;
 
 @Service
 public class ProductServiceImpl implements ProductService{
@@ -23,69 +25,76 @@ public class ProductServiceImpl implements ProductService{
 	private CategoryDao cdao;
 	
 	@Autowired
-	private CategoryService cservice;
+	private SessionDao sdao; 
 	
 	@Override
-	public List<Product> viewAllProduct() throws ProductException {
+	public List<Product> viewAllProducts() throws ProductException {
 		
-		List<Product> allProduct = prepo.findAll();
+		List<Product> allProducts = prepo.findAll();
 		
-		if(allProduct.size() == 0) {
+		if(allProducts.size() == 0) {
 			throw new ProductException("Product does not exists...");
 		}
 		else {
-			return allProduct;
+			return allProducts;
 		}
 		
 	}
 
 	@Override
-	public Product addProduct(Product product) throws ProductException {
+	public Product addProduct(Product product, String key) throws ProductException {
 		
-		Product existingProduct = prepo.findByProductName(product.getProductName());
-		
-		if(existingProduct == null) {
+		CurrentSession validSession = sdao.findByUuid(key);
+		if(validSession!=null && validSession.getRole().toString().equals("ADMIN")) {
+			Product existingProduct = prepo.findByProductName(product.getProductName());
 			
-			return prepo.save(product);
-			
-		}
-		else {
-			
-			if(existingProduct.getManufacturer().equals(product.getManufacturer())) {
-				
-				throw new ProductException("Product already exixts...");
-				
-			}
-			else {
+			if(existingProduct == null) {
 				
 				return prepo.save(product);
 				
 			}
-			
+			else {
+				
+				if(existingProduct.getManufacturer().equals(product.getManufacturer())) {
+					
+					throw new ProductException("Product already exixts...");
+					
+				}
+				else {
+					
+					return prepo.save(product);
+					
+				}
+				
+			}
 		}
+		throw new ProductException("Either invalid key or not of an Admin.");
 		
 	}
 
 	@Override
-	public Product updateProduct(Product product) throws ProductException {
-		
-		Product existingProduct = prepo.findByProductName(product.getProductName());
-		
-		if(existingProduct == null) {
-			throw new ProductException(product.getProductName()+" does not have in the product list");
+	public Product updateProduct(Product product, String key) throws ProductException {
+		CurrentSession validSession = sdao.findByUuid(key);
+		if(validSession!=null && validSession.getRole().toString().equals("ADMIN")) {
+			Product existingProduct = prepo.findByProductName(product.getProductName());
+			
+			if(existingProduct == null) {
+				throw new ProductException(product.getProductName()+" does not have in the product list");
+			}
+			else {
+				existingProduct.setColor(product.getColor());
+				existingProduct.setManufacturer(product.getManufacturer());
+				existingProduct.setDimension(product.getDimension());
+				existingProduct.setPrice(product.getPrice());
+				existingProduct.setProductName(product.getProductName());
+				existingProduct.setSpecification(product.getSpecification());
+				
+				Product updatedProduct = prepo.save(existingProduct);
+				
+				return updatedProduct;
+			}
 		}
-		else {
-			
-			existingProduct.setColor(product.getColor());
-			existingProduct.setDimension(product.getDimension());
-			existingProduct.setManufacturer(product.getManufacturer());
-			existingProduct.setPrice(product.getPrice());
-			existingProduct.setProductName(product.getProductName());
-			existingProduct.setSpecification(product.getSpecification());
-			
-			return prepo.save(existingProduct);
-			
-		}
+		throw new ProductException("Either invalid key or not of an Admin.");
 	}
 
 	@Override
@@ -103,7 +112,7 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public List<Product> viewProductByCategory(String categoryName) throws ProductException, CategoryException {
+	public List<Product> viewProductsByCategory(String categoryName) throws ProductException, CategoryException {
 		
 		Category existingCategory = cdao.findByCategoryName(categoryName);
 		
@@ -132,51 +141,58 @@ public class ProductServiceImpl implements ProductService{
 	}
 
 	@Override
-	public Product removeProduct(Integer productID) throws ProductException {
-		
-		Optional<Product> existingProduct = prepo.findById(productID);
-		
-		if(existingProduct.isEmpty()) {
-			throw new ProductException("Product not found...");
+	public Product removeProduct(Integer productID, String key) throws ProductException {
+		CurrentSession validSession = sdao.findByUuid(key);
+		if(validSession!=null && validSession.getRole().toString().equals("ADMIN")) {
+			Optional<Product> existingProduct = prepo.findById(productID);
+			
+			if(existingProduct.isEmpty()) {
+				throw new ProductException("Product not found...");
+			}
+			else {
+				
+				Product product = existingProduct.get();
+				
+				prepo.delete(product);
+				
+				return product;
+			}
 		}
-		else {
-			
-			Product product = existingProduct.get();
-			
-			prepo.delete(product);
-			
-			return existingProduct.get();
-		}
+		throw new ProductException("Either invalid key or not of an Admin.");
 		
 	}
 
 	@Override
-	public Product addCategoryToTheProduct(Integer productId, String cname) throws ProductException, CategoryException {
+	public Product addCategoryToTheProduct(Integer productId, String cname, String key) throws ProductException, CategoryException {
 		
-		Category category = cdao.findByCategoryName(cname);
-		
-		if(category == null) {
-			throw new CategoryException("Category does not exists please add this category...");
-		}
-		else {
+		CurrentSession validSession = sdao.findByUuid(key);
+		if(validSession!=null && validSession.getRole().toString().equals("ADMIN")) {
+			Category category = cdao.findByCategoryName(cname);
 			
-			Optional<Product> product = prepo.findById(productId);
-			
-			if(product.isPresent()) {
-				
-				Product existingProduct = product.get();
-				
-				existingProduct.setCategory(category);
-				
-				return prepo.save(existingProduct);
-				
+			if(category == null) {
+				throw new CategoryException("Category does not exists please add this category...");
 			}
 			else {
-				throw new ProductException("Product does not exists...");
+				
+				Optional<Product> product = prepo.findById(productId);
+				
+				if(product.isPresent()) {
+					
+					Product existingProduct = product.get();
+					
+					existingProduct.setCategory(category);
+					
+					return prepo.save(existingProduct);
+					
+				}
+				else {
+					throw new ProductException("Product does not exists...");
+				}
+				
 			}
-			
 		}
 		
+		throw new ProductException("Either invalid key or not of an Admin.");
 	}
 	
 }
